+++ lib/sprite_animation.dart
@@ -1,0 +1,1 @@
+export 'src/sprite_animation.dart';