+++ lib/sprite.dart
@@ -1,0 +1,4 @@
+export 'src/sprite.dart';
+export 'src/sprite_animation.dart';
+export 'src/sprite_batch.dart';
+export 'src/spritesheet.dart';