+++ lib/sprite_batch.dart
@@ -1,1 +1,0 @@
-export 'src/sprite_batch.dart';