+++ doc/examples/parallax/lib/main_game.dart
@@ -2,8 +2,14 @@ import 'package:flame/components.dart';
 import 'package:flame/flame.dart';
 import 'package:flame/game.dart';
 import 'package:flame/parallax.dart';
+import 'package:flame/extensions.dart';
 import 'package:flutter/material.dart';
 
+/// This examples serves to test the Parallax feature outside of the
+/// Flame Component System (FCS), use the other files in this folder
+/// for examples on how to use parallax with FCS
+/// FCS is only used when you extend BaseGame, not Game,
+/// like we do in this example.
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
   await Flame.device.fullScreen();
@@ -14,17 +20,12 @@ void main() async {
   );
 }
 
-class MyGame extends BaseGame {
-  @override
-  Future<void> onLoad() async {
-    add(MyParallaxComponent());
-  }
-}
+class MyGame extends Game {
+  Parallax parallax;
 
-class MyParallaxComponent extends ParallaxComponent with HasGameRef<MyGame> {
   @override
   Future<void> onLoad() async {
-    parallax = await gameRef.loadParallax(
+    parallax = await loadParallax(
       [
         'bg.png',
         'mountain-far.png',
@@ -36,5 +37,15 @@ class MyParallaxComponent extends ParallaxComponent with HasGameRef<MyGame> {
       baseVelocity: Vector2(20, 0),
       velocityMultiplierDelta: Vector2(1.8, 1.0),
     );
+  }
+
+  @override
+  void update(double dt) {
+    parallax.update(dt);
+  }
+
+  @override
+  void render(Canvas canvas) {
+    parallax.render(canvas);
   }
 }