+++ lib/src/components/joystick/joystick_component.dart
@@ -1,9 +1,7 @@
-import 'dart:ui';
+import 'package:meta/meta.dart';
 
-import '../../extensions/vector2.dart';
+import '../../../components.dart';
 import '../../game/base_game.dart';
-import '../../gestures.dart';
-import '../component.dart';
 import '../mixins/has_game_ref.dart';
 import 'joystick_action.dart';
 import 'joystick_directional.dart';
@@ -14,7 +12,8 @@ mixin JoystickListener {
   void joystickAction(JoystickActionEvent event);
 }
 
-abstract class JoystickController extends Component with HasGameRef<BaseGame> {
+abstract class JoystickController extends BaseComponent
+    with HasGameRef<BaseGame>, Draggable {
   final List<JoystickListener> _observers = [];
 
   void joystickChangeDirectional(JoystickDirectionalEvent event) {
@@ -29,58 +28,35 @@ abstract class JoystickController extends Component with HasGameRef<BaseGame> {
     _observers.add(listener);
   }
 
-  void onReceiveDrag(DragEvent drag) {}
-
   @override
   bool isHud = true;
 }
 
 class JoystickComponent extends JoystickController {
-  final List<JoystickAction> actions;
-  final JoystickDirectional directional;
   @override
   int priority;
 
   JoystickComponent({
-    this.actions,
-    this.directional,
+    @required BaseGame gameRef,
+    List<JoystickAction> actions = const [],
+    JoystickDirectional directional,
     this.priority = 0,
-  });
+  }) {
+    if (directional != null) {
+      addChild(directional, gameRef: gameRef);
+    }
+    actions.forEach((action) => addChild(action, gameRef: gameRef));
+  }
 
   void addAction(JoystickAction action) {
-    if (gameRef?.size != null) {
-      action.initialize(gameRef.size, this);
-      actions?.add(action);
-    }
+    addChild(action, gameRef: gameRef);
   }
 
   void removeAction(int actionId) {
-    actions?.removeWhere((action) => action.actionId == actionId);
-  }
-
-  @override
-  void render(Canvas canvas) {
-    directional?.render(canvas);
-    actions?.forEach((action) => action.render(canvas));
-  }
-
-  @override
-  void update(double t) {
-    super.update(t);
-    directional?.update(t);
-    actions?.forEach((action) => action.update(t));
-  }
-
-  @override
-  void onGameResize(Vector2 size) {
-    directional?.initialize(size, this);
-    actions?.forEach((action) => action.initialize(size, this));
-    super.onGameResize(size);
-  }
-
-  @override
-  void onReceiveDrag(DragEvent event) {
-    directional?.onReceiveDrag(event);
-    actions?.forEach((action) => action.onReceiveDrag(event));
+    final action = children
+        .firstWhere((e) => e is JoystickAction && e.actionId == actionId);
+    if (action != null) {
+      removeChild(action);
+    }
   }
 }