+++ CHANGELOG.md
@@ -9,6 +9,8 @@
  - Enabling direct import of Sprite and SpriteAnimation
  - Renamed `Composition` to `ImageComposition` to prevent confusion with the composition component
  - Added `rotation` and `anchor` arguments to `ImageComposition.add`
+ - Added `Image` extensions
+ - Added `Color` extensions
  - Change RaisedButton to ElevatedButton in timer example
 
 ## 1.0.0-rc6